+++ SortColumnsOfMatrixByFirstElements/SortColumnsOfMatrixByFirstElements/Program.cs
@@ -1,11 +1,10 @@
-﻿using System;
-
-namespace SortColumnsOfMatrixByFirstElements
+﻿namespace SortColumnsOfMatrixByFirstElements
 {
     class Program
     {
         static void Main(string[] args)
         {
+            Task.UserInteraface();
         }
     }
 }
