+++ SortColumnsOfMatrixByFirstElements/SortColumnsOfMatrixByFirstElements/SortColumnsOfMatrixByFirstElements.csproj
@@ -43,8 +43,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
-    <Compile Include="Class2.cs" />
+    <Compile Include="Sort.cs" />
+    <Compile Include="Task.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
